src/final_requirements.txt
arxiv==2.1.3
beautifulsoup4==4.12.3
bs4==0.0.2
cassio==0.1.10
faiss-cpu==1.9.0
fastapi==0.115.4
groq==0.11.0
langchain==0.1.20
langchain-astradb==0.4.0
langchain-community==0.0.38
langchain-core==0.1.53
langchain-experimental==0.0.58
langchain-groq==0.1.5
langchain-objectbox==0.1.0
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langchainhub==0.1.21
langserve==0.2.3
pypdf==5.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.39.0
uvicorn==0.32.0
wikipedia==1.4.0